file renamed to send-verification-email.js
